runtime: Make runtime a separately compiled static library
