+++ cmake/install-rules.cmake
@@ -9,10 +9,10 @@ set(Jakt_INSTALL_CMAKEDIR "${CMAKE_INSTALL_DATADIR}/Jakt"
     CACHE PATH "CMake package config location relative to the install prefix")
 mark_as_advanced(Jakt_INSTALL_CMAKEDIR)
 
+configure_file(cmake/install-config.cmake.in JaktConfig.cmake @ONLY)
 install(
-    FILES cmake/install-config.cmake
+    FILES "${CMAKE_CURRENT_BINARY_DIR}/JaktConfig.cmake"
     DESTINATION "${Jakt_INSTALL_CMAKEDIR}"
-    RENAME JaktConfig.cmake
     COMPONENT Jakt_Development
 )
 
@@ -44,6 +44,21 @@ foreach (stage IN LISTS stages)
   )
 endforeach()
 
+install(
+    TARGETS jakt_runtime
+    EXPORT JaktTargets
+    RUNTIME #
+        DESTINATION "${CMAKE_INSTALL_BINDIR}"
+        COMPONENT Jakt_Runtime
+    LIBRARY #
+        DESTINATION "${CMAKE_INSTALL_LIBDIR}"
+        COMPONENT Jakt_Runtime
+        NAMELINK_COMPONENT Jakt_Development
+    ARCHIVE #
+        DESTINATION "${CMAKE_INSTALL_LIBDIR}"
+        COMPONENT Jakt_Development
+)
+
 # Install runtime files. In the future we should probably have a bazillion rules
 #   for compiling this into a static archive for different configs
 install(DIRECTORY "runtime"
@@ -54,8 +69,6 @@ install(DIRECTORY "runtime"
                        PATTERN "utility"
 )
 
-# FIXME: Figure out how to get an autogenerated import target for the final compiler (Jakt::jakt), 
-#        which is the symlink below
 install(
     EXPORT JaktTargets
     NAMESPACE Jakt::
